Basic Functionality
I did this last night but didn't have time to upload to github.

This is not finished.

Managed to use the EMBY API to find media on my server and get TMDB ID.

I then use TMDB to search for the trailer URL using the ID from EMBY.

At the moment some things are hard coded for now.  So I will be cleaning that up in a future release.