Adjust Environment Variables
I have decided to remove the old way of specifying the protocol, host and port manually as it just adds to confusion.
I am now just having the single key called host that will contrain all the relevent information for sending the request.
This makes error checking a lot easier.